+++ podchat/src/main/java/com/fanap/podchat/mainmodel/ChatMessage.java
@@ -3,11 +3,11 @@ package com.fanap.podchat.mainmodel;
 public class ChatMessage extends AsyncMessage {
 
     private Integer messageType;
-    private long time;
-    private int contentCount;
+    private Long time;
+    private Integer contentCount;
     private String systemMetadata;
     private String metadata;
-    private long repliedTo;
+    private Long repliedTo;
 
 
     public long getTime() {